chore: update commit, push steps
* first change was moving the commit step closer to the push step in
the overall workflow. Now, branch protection happens first, then commit,
then push. This should reduce the time between commit and push.
* second change was adding a pull command before changes are committed.
This should tighten the window of this step failing because a change
to main is pushed and not pulled.
* the last change is grouping the echo command in the commit step with
the commit command. Now, echo setting the output should only happen when
we have changes to commit. If we dont have any changes, echo shouldnt
occur, and the push step should not run.