Update and rename mssql-queryplan.json to mssql-customqueries.json
This change is due to the fact that this quickstart was initially created from an experimental integration that sends information about Query Plans in Logs. However, in this dashboard these Query Plans Logs are not consulted anywhere. I have found that this is generating a lot of confusion among users who expect that after enabling this integration these widgets will become popular.
The reality is that this dashboard is mainly powered by the basic MSSQL integration + the custom SQL queries that are executed in this file.