fix: Set to actual ID from database
This is due to this config being merged into main and released. The Catalog Service assigned an ID of all 0's because our system had not generated an ID for it yet. Leaving it in as this ID because it is valid and to prevent a similar issue from occurring in the future.

Co-authored-by: Alec Swanson <aswanson@newrelic.com>